pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_vertexai_search4commerce"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp==2.10.4",
    "python-dotenv",
    "google-cloud-retail",
]


